Few css corrections

+++ views/pages/userProfile.php
@@ -17,13 +17,38 @@ $ownerAccount = $_GET['action'] === 'myProfile';// Check if it's the owner accou
 
             <hr />
             <h2 class="playfair-font"><?= $user->getUsername() ?></h2>
-            <p>Inscrit depuis : <?= Utils::getRegistrationDuration($user->getCreatedAt()) ?></p>
-            <div>
-                <h3>Bibliothèque</h3>
+            <p class="registration_duration">Membre depuis : <?= Utils::getRegistrationDuration($user->getCreatedAt()) ?></p>
+            <div class="librairy_count">
+                <h3 class="font-semibold">Bibliothèque</h3>
                 <p>
                     <img src="./assets/librairy_icon.svg" alt="" width="10" height="13">
-                    <?= count($library) ?> livres</p>
+                    <?= count($library) ?> livres
+                </p>
             </div>
+            <?php if (!$ownerAccount) {
+                ?>
+                <a href="index.php?action=openChat&idReceiver=<?= $user ? $user->getId() : null ?>"
+                    class="secondary_button font-semibold primary_button--full_width font-semibold">
+                    Écrire un message
+                </a>
+                <button id="modale_opener" class="hidden secondary_button font-semibold primary_button--full_width font-semibold">
+                    Écrire un message
+                </button>
+                <dialog>
+                    <div class="dialog_container">
+                        <button id="modale_closer">X</button>
+                        <form action="index.php?action=sendMessageWithAjax" method="post">
+                            <label for="message">Envoyer un message à <?= $user ? $user->getUsername() : null ?></label>
+                            <textarea autofocus name="message" id="message" cols="30" rows="10"></textarea>
+                            <button type="dialog" data-idReceiver=<?= $user ? $user->getId() : null ?>
+                                class="primary_button primary_button--full_width font-semibold">Confirmation</button>
+                        </form>
+                    </div>
+                    </p>
+                </dialog>
+                <?php
+            } ?>
+
         </div>
         <?php
         if ($ownerAccount) { ?>
@@ -49,34 +74,39 @@ $ownerAccount = $_GET['action'] === 'myProfile';// Check if it's the owner accou
             <?php
         } ?>
     </section>
-
     <section class="account_librairy<?= !$ownerAccount ? ' account_librairy--smaller' : '' ?>">
         <a href="index.php?action=newBookForm">Ajouter un livre.</a>
-        <div class="first_line">
-            <div>Photo</div>
-            <div>Titre</div>
-            <div>Auteur</div>
-            <div>Description</div>
-            <div>Disponibilité</div>
-            <div>Action</div>
-        </div>
-        <?php
-        foreach ($library as $book) {
-            ?>
-            <div class="book_row">
-                <img src=<?= $book ? Utils::filepath($book->getFilename()) : null ?> alt="book cover" width="78" height="78">
-                <h2><?= $book->getTitle() ?></h2>
-                <h2><?= $book->getAuthor() ?></h2>
-                <p><?= $book->getDescription() ?></p>
-                <p><?= $book->getAvailability() ? 'Disponible' : 'Indisponible' ?></p>
-                <div>
-                    <a href="index.php?action=deleteBook&id=<?= $book->getId() ?>">Supprimer</a>
-                    <a href="index.php?action=editBookForm&id=<?= $book->getId() ?>">Éditer</a>
-                </div>
-            </div>
-            <?php
-        }
-        ?>
-    </section>
+        <table>
+            <thead>
+                <tr>
+                    <th>Photo</th>
+                    <th>Titre</th>
+                    <th>Auteur</th>
+                    <th>Description</th>
+                    <th>Disponibilité</th>
+                    <?php if ($ownerAccount) { ?>
+                        <th>Action</th>
+                    <?php } ?>
+                </tr>
+            </thead>
+            <tbody>
+                <?php foreach ($library as $book) { ?>
+                    <tr>
+                        <td><img src=<?= $book ? Utils::filepath($book->getFilename()) : null ?> alt="book cover" width="78" height="78"></td>
+                        <td><?= $book->getTitle() ?></td>
+                        <td><?= $book->getAuthor() ?></td>
+                        <td><div><?= $book->getDescription() ?></div></td>
+                        <td><?= $book->getAvailability() ? 'Disponible' : 'Indisponible' ?></td>
+                        <?php if ($ownerAccount) { ?>
+                            <td>
+                                <a href="index.php?action=deleteBook&id=<?= $book->getId() ?>">Supprimer</a>
+                                <a href="index.php?action=editBookForm&id=<?= $book->getId() ?>">Éditer</a>
+                            </td>
+                        <?php } ?>
+                    </tr>
+                <?php } ?>
+            </tbody>
+        </table>
 
+    </section>
 </section>
